+++ src/banner/render/IMinFPSLegacyProps.ts
@@ -1,9 +1,11 @@
-export interface IMinFPSLegacyProps {
-  // [key: string]: string | boolean ;
-    //Section 16 - FPS Options group
-    searchShow: boolean;
-    fpsPageStyle: string;
-    fpsContainerMaxWidth: string;
-}
+// NO LONGER NEEDED
 
-export const changefpsOptions2 : string[] = [  'fpsPageStyle', 'fpsContainerMaxWidth' ];
+// export interface IMinFPSLegacyProps {
+//   // [key: string]: string | boolean ;
+//     //Section 16 - FPS Options group
+//     searchShow: boolean;
+//     fpsPageStyle: string;
+//     fpsContainerMaxWidth: string;
+// }
+
+// export const changefpsOptions2 : string[] = [  'fpsPageStyle', 'fpsContainerMaxWidth' ];